moved app.js to pages folder+updated accordingly in index.js
